fix(runtime): deduplicate host-fair worktree indexes
